+++ src/main/java/crying/interfaces/SanityManager.java
@@ -10,7 +10,6 @@ import net.minecraft.entity.LivingEntity;
 import net.minecraft.nbt.NbtCompound;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.server.world.ServerWorld;
-import net.minecraft.sound.SoundCategory;
 import net.minecraft.sound.SoundEvents;
 import net.minecraft.world.Difficulty;
 import net.minecraft.world.GameRules;
@@ -214,7 +213,7 @@ public class SanityManager {
         updateThis();
 
         if (entity != null) {
-            entity.getWorld().playSound(entity, entity.getX(), entity.getY(), entity.getZ(), SoundEvents.ENTITY_PLAYER_HURT, SoundCategory.PLAYERS);
+            entity.getWorld().playSound(null, entity.getX(), entity.getY(), entity.getZ(), SoundEvents.ENTITY_PLAYER_HURT, entity.getSoundCategory());
         }
     }